+++ NEWS.md
@@ -1,3 +1,5 @@
+* Added the function `add_zenodo_citation`.
+
 # add2pkg 0.1.0
 
 * Added the function `add_me`.